Dispose MarkdownReader
